Bootstrap measures

Bootstrap

TextStylometry.equal_windows — Function.

equal_windows(N,window)

Create interval sequence given a length N and window window.

The last window chunk is not returned if leftover is less than window.

source

TextStylometry.bootstrap_window — Function.

bootstrap_window(f,tokens,window)

Calculate f on window-size chunks of tokens.

source

TextStylometry.bootstrap_ci — Function.

bootstrap_ci(xs)

Calculates a 95% confidence interval given bootstrap sequence xs.

source

TextStylometry.sttr — Function.

sttr(tokens,window)

Calculate the Standardized Type-Token Ratio on tokens given window size.

source